requirement.txt
absl-py==2.3.1
aiodns==3.3.0
aiofiles==25.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-cors==0.7.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
archspec==0.2.5
argcomplete==3.6.3
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
Beaker==1.12.1
beautifulsoup4==4.14.2
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
boltons==25.0.0
boto3==1.40.72
botocore==1.40.72
Brlapi==0.8.6
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
comm==0.2.3
conda==24.11.3
conda-libmamba-solver==24.9.0
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnf==4.24.0
executing==2.2.1
fastapi==0.122.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.20.0
Flask==3.1.1
flatbuffers==25.9.23
fros==1.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.10.0
gast==0.6.0
gbinder-python==1.1.2
git-filter-repo==2.47.0
gns3-gui==2.2.54
gns3-server==2.2.54
google-pasta==0.2.0
grpcio==1.76.0
h11==0.14.0
h5py==3.15.1
httpcore==1.0.7
httpx==0.27.2
humanize==4.12.0
idna==3.10
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
iso639==0.1.4
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keras==3.12.0
langtable==0.0.69
libclang==18.1.1
libcomps==0.1.22
libdnf==0.75.0
libmambapy==1.5.12
libvirt-python==11.0.0
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mdurl==0.1.2
menuinst==2.0.0
mercurial==6.9.5
ml_dtypes==0.5.4
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.6
nftables==0.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
olefile==0.47
opt_einsum==3.4.0
optree==0.18.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
productmd==1.45
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
py-cpuinfo==9.0.0
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycairo==1.25.1
pycares==4.10.0
pyclip==0.7.0
pycosat==0.6.6
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.12.5
pydantic_core==2.41.5
pyenchant==3.2.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
PyMsgBox==1.0.9
pyOpenSSL==25.0.0
pyparted==3.13.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5_sip==12.16.1
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
pyudev==0.24.3
pywhatkit==5.4
pyxdg==0.27
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
rich==14.2.0
rpds-py==0.25.0
rpm==4.20.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.12
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.3
setools==4.5.1
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
systemd-python==235
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ufw==0.35
urllib3==2.3.0
uvicorn==0.35.0
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==2.0.1
xgboost==3.1.2
xkbregistry==0.3
yarl==1.13.1
zstandard==0.23.0
